requirements.txt
selenium
requests
suds-jurko
lxml
pillow
numpy
pyyaml
python-box